more compile corrections

+++ src/runtime_src/xdp/profile/plugin/aie_debug/aie_debug_impl.h
@@ -39,7 +39,7 @@ namespace xdp {
 
     virtual void updateDevice() = 0;
     virtual void updateAIEDevice(void* handle)=0;
-    virtual void poll(const uint32_t index, void* handle) = 0;
+    virtual void poll(const uint64_t index, void* handle) = 0;
 
     const std::map<module_type, const char*> moduleTypes = {
       {module_type::core, "AIE"},
