+++ src/runtime_src/xdp/profile/plugin/aie_debug/edge/aie_debug.h
@@ -30,7 +30,7 @@ namespace xdp {
     ~AieDebug_EdgeImpl();
     void updateDevice();
     void updateAIEDevice(void* handle);
-    void poll(const uint32_t index, void* handle);
+    void poll(const uint64_t index, void* handle);
 
   private:
